backend/pyproject.toml
[tool.poetry]
name = "flowbeat-backend"
version = "0.1.0"
description = "FlowBeat Music Streaming & Recommendation System Backend"
authors = ["FlowBeat Team"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"

# =============================================================================
# Web 框架核心
# =============================================================================
# 使用 FastAPI 构建高性能异步 API
fastapi = "^0.109.0"
# ASGI 服务器: Uvicorn 是生产级的 ASGI 服务器
uvicorn = { extras = ["standard"], version = "^0.27.0" }
# [关键依赖] 多部分解析: OAuth2 的表单登录 (application/x-www-form-urlencoded) 强依赖此库
# 如果缺失，登录接口会报 500 错误
python-multipart = "^0.0.9"
# 邮箱验证: 这里的 v2.1.0 提供了更严格的 RFC 校验
email-validator = "^2.1.0"

# =============================================================================
# 数据库与 ORM (全链路异步)
# =============================================================================
# SQLAlchemy 2.0+ 原生支持 async/await
sqlalchemy = { extras = ["asyncio"], version = "^2.0.25" }
# 数据库迁移工具
alembic = "^1.13.1"
# PostgreSQL 异步驱动，性能远超 psycopg2
asyncpg = "^0.29.0"

# =============================================================================
# 配置与数据校验
# =============================================================================
# Pydantic V2 提供了极强的性能和类型安全性
pydantic = "^2.6.0"
pydantic-settings = "^2.1.0"

# =============================================================================
# 认证与安全 (Auth & Security)
# =============================================================================
# Python-jose 处理 JWT 编解码 (生成 Bearer Token)
python-jose = { extras = ["cryptography"], version = "^3.3.0" }

# [架构升级] 密码哈希算法 (Argon2id)
# 1. 移除 extras=["bcrypt"] 以解决 Windows 环境下的编译与依赖冲突
# 2. 启用 extras=["argon2"] 以适配现代安全标准
passlib = { extras = ["argon2"], version = "^1.7.4" }

# [核心依赖] Argon2 CFFI 绑定库
# 显式声明版本，确保生产环境使用稳定的 Argon2id 实现，防止 passlib 找不到后端
argon2-cffi = "^23.1.0"

# =============================================================================
# 工具库
# =============================================================================
httpx = "^0.26.0"
python-slugify = "^8.0.1"

[tool.poetry.group.dev.dependencies]
# 测试框架
pytest = "^8.0.0"
pytest-asyncio = "^0.23.5"
# 代码格式化与 Lint
black = "^24.1.1"
isort = "^5.13.2"
pre-commit = "^3.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"